+++ tensorflow/core/platform/status.cc
@@ -74,7 +74,7 @@ class StatusLogSink : public TFLogSink {
 
     mutex_lock lock(mu_);
     messages_.emplace_back(entry.ToString());
-    if (messages_.size() > num_messages_) messages_.pop_front();
+    if (messages_.size() > static_cast<size_t>(num_messages_)) messages_.pop_front();
   }
 
  private:
